+++ packages/Flow/src/package-test.ts
@@ -20,6 +20,7 @@ import './tests/files-tree-tests';
 import './tests/execution-preview-tests';
 import './tests/viewer-tests';
 import './tests/column-picker-tests';
+import './tests/connect-interaction-tests';
 
 export let _package = new DG.Package();
 export { tests };